# backend/openapi-schema.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "MbPA Building Permission Portal API",
        "version": "0.1.0",
        "description": "Mumbai Port Authority Special Planning Authority — UPDR-2026 building permission workflow."
    },
    "paths": {
...
        "/api/compliance/complaints/": {
            "get": {
                "operationId": "complaints_list",
                "tags": [
                    "compliance"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ComplaintRead"
                                    }
                                }
                            }
                        },
                        "description": ""
                    }
                }
            },
            "post": {
                "operationId": "compliance_complaints_create",
                "tags": [
                    "compliance"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ComplaintCreateRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/ComplaintCreateRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ComplaintCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ComplaintRead"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/compliance/complaints/{id}/": {
            "get": {
                "operationId": "complaints_retrieve",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "compliance"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ComplaintRead"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            },
            "patch": {
                "operationId": "compliance_complaints_partial_update",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "compliance"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedComplaintResolveRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedComplaintResolveRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedComplaintResolveRequest"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ComplaintRead"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/compliance/erasure-requests/": {
            "get": {
                "operationId": "erasure_requests_list",
                "description": "AC-32: DPDP erasure requests. Anyone may request erasure of their own data.",
                "tags": [
                    "compliance"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ErasureRequestRead"
                                    }
                                }
                            }
                        },
                        "description": ""
                    }
                }
            },
            "post": {
                "operationId": "compliance_erasure_requests_create",
                "description": "AC-32: DPDP erasure requests. Anyone may request erasure of their own data.",
                "tags": [
                    "compliance"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ErasureRequestCreateRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/ErasureRequestCreateRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ErasureRequestCreateRequest"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErasureRequestRead"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ComplaintCreateRequest": {
                "type": "object",
                "properties": {
                    "application_id": {
                        "type": "integer"
                    },
                    "subject": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 255
                    },
                    "body": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "application_id",
                    "body",
                    "subject"
                ]
            },
            "ComplaintRead": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "application": {
                        "type": "integer"
                    },
                    "origin": {
                        "$ref": "#/components/schemas/OriginEnum"
                    },
                    "raised_by": {
                        "type": "integer",
                        "nullable": true
                    },
                    "subject": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "body": {
                        "type": "string"
                    },
                    "status": {
                        "$ref": "#/components/schemas/ComplaintReadStatusEnum"
                    },
                    "resolution_notes": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "resolved_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "required": [
                    "application",
                    "body",
                    "created_at",
                    "id",
                    "origin",
                    "subject",
                    "updated_at"
                ]
            },
            "ComplaintReadStatusEnum": {
                "enum": [
                    "open",
                    "in_review",
                    "resolved",
                    "closed"
                ],
                "type": "string",
                "description": "* `open` - Open\n* `in_review` - In Review\n* `resolved` - Resolved\n* `closed` - Closed"
            },
...
            "ErasureRequestCreateRequest": {
                "type": "object",
                "properties": {
                    "subject_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "reason": {
                        "type": "string",
                        "default": ""
                    }
                }
            },
            "ErasureRequestRead": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "subject": {
                        "type": "integer"
                    },
                    "requested_by": {
                        "type": "integer",
                        "nullable": true
                    },
                    "reason": {
                        "type": "string"
                    },
                    "status": {
                        "$ref": "#/components/schemas/ErasureRequestReadStatusEnum"
                    },
                    "requested_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "due_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "processed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "processed_by": {
                        "type": "integer",
                        "nullable": true
                    },
                    "resolution_notes": {
                        "type": "string"
                    },
                    "is_overdue": {
                        "type": "boolean",
                        "readOnly": true
                    }
                },
                "required": [
                    "due_at",
                    "id",
                    "is_overdue",
                    "requested_at",
                    "subject"
                ]
            },
            "ErasureRequestReadStatusEnum": {
                "enum": [
                    "pending",
                    "completed",
                    "rejected"
                ],
                "type": "string",
                "description": "* `pending` - Pending\n* `completed` - Completed\n* `rejected` - Rejected"
            },
...
            "OriginEnum": {
                "enum": [
                    "applicant_raised",
                    "system_raised"
                ],
                "type": "string",
                "description": "* `applicant_raised` - Applicant Raised\n* `system_raised` - System Raised"
            },
...
            "PatchedComplaintResolveRequest": {
                "type": "object",
                "properties": {
                    "resolution_notes": {
                        "type": "string",
                        "minLength": 1
                    }
                }
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            }
        }
    }
}